Warning level 4 on retrofit, warnings disabled on gtest

+++ apps/retrofit/source/project.h
@@ -2,7 +2,7 @@
 #include "Logging/Constants.h"
 
 constexpr char const* RFCAT_STARTUP = "Startup";
-constexpr uint64_t RFCAT_STARTUP_SEV_WHITELIST = -1;
+constexpr uint64_t RFCAT_STARTUP_SEV_WHITELIST = static_cast<uint64_t>( -1 );
 
 constexpr char const* RFCAT_FRAMEPACKEDITOR = "FramePackEditor";
-constexpr uint64_t RFCAT_FRAMEPACKEDITOR_SEV_WHITELIST = -1;
+constexpr uint64_t RFCAT_FRAMEPACKEDITOR_SEV_WHITELIST = static_cast<uint64_t>( -1 );
